+++ src/index.js
@@ -11,12 +11,14 @@ const getAsNumber = require("./getAsNumber.js");
 const getAsObject = require("./getAsObject.js");
 const getAsString = require("./getAsString.js");
 const groupBy = require("./groupBy.js");
+const isNonEmptyArray = require("./isNonEmptyArray.js");
 const isNonEmptyObject = require("./isNonEmptyObject.js");
 const isNullOrUndefined = require("./isNullOrUndefined.js");
 const isObject = require("./isObject.js");
 const last = require("./last.js");
 const pick = require("./pick.js");
 const range = require("./range.js");
+const sortBy = require("./sortBy.js");
 const toCamelCase = require("./toCamelCase.js");
 const uniqBy = require("./uniqBy.js");
 const uniqueId = require("./uniqueId.js");
@@ -36,12 +38,14 @@ module.exports = {
 	getAsObject,
 	getAsString,
 	groupBy,
+	isNonEmptyArray,
 	isNonEmptyObject,
 	isNullOrUndefined,
 	isObject,
 	last,
 	pick,
 	range,
+	sortBy,
 	toCamelCase,
 	uniqBy,
 	uniqueId,