+++ src/isNonEmptyArray.js
@@ -1,0 +1,5 @@
+module.exports = function isNonEmptyArray (arr) {
+	return Boolean(
+		Array.isArray(arr) && arr.length
+	);
+};